[add] first iteration on ingestion with hashes
