+++ cmd/ftsb_load_redisearch/main.go
@@ -6,13 +6,9 @@ import (
 	"fmt"
 	"github.com/RediSearch/ftsb/load"
 	"github.com/RediSearch/redisearch-go/redisearch"
+	//"github.com/mediocregopher/radix"
 	"log"
-	"math/rand"
-	"os"
-	"strconv"
 	"strings"
-	"sync"
-	"time"
 )
 
 // Program option vars:
@@ -27,6 +23,8 @@ var (
 	syntheticsNumberFields  uint64
 	loader                  *load.BenchmarkRunner
 	useCase                 string
+	useHmset                bool
+	useFtadd                bool
 )
 
 const (
@@ -72,7 +70,8 @@ func init() {
 	flag.Uint64Var(&syntheticsCardinality, "synthetic-max-dataset-cardinality", 1024, "Max Field cardinality specific to the synthetics use cases (e.g., distinct tags in 'tag' fields).")
 	flag.Uint64Var(&syntheticsNumberFields, "synthetic-fields", 10, "Number of fields per document specific to the synthetics use cases (starting at field1, field2, field3, etc...).")
 	flag.StringVar(&useCase, "use-case", "enwiki-abstract", fmt.Sprintf("Use case to model. (choices: %s)", strings.Join(useCaseChoices, ", ")))
-
+	flag.BoolVar(&useHmset, "use-hmset", false, "If set to true, it will use hmset command to insert the documents.")
+	flag.BoolVar(&useFtadd, "use-ftadd", false, "If set to true, it will use ft.add to insert the documents.")
 	flag.IntVar(&debug, "debug", 0, "Debug printing (choices: 0, 1, 2). (default 0)")
 	flag.Parse()
 }
@@ -115,158 +114,11 @@ func (b *benchmark) GetPointIndexer(maxPartitions uint) load.PointIndexer {
 }
 
 func (b *benchmark) GetProcessor() load.Processor {
-	return &processor{b.dbc, nil, nil, nil, nil, nil, nil, nil, nil, []string{}, []string{}, []string{},}
+	return &processor{b.dbc, nil, nil, nil, nil, nil, nil, nil, nil, []string{}, []string{}, []string{}, nil}
 }
 
 func (b *benchmark) GetDBCreator() load.DBCreator {
 	return b.dbc
-}
-
-type processor struct {
-	dbc              *dbCreator
-	rows             chan string
-	insertsChan      chan uint64
-	totalLatencyChan chan uint64
-	updatesChan      chan uint64
-	deletesChan      chan uint64
-	totalBytesChan   chan uint64
-	wg               *sync.WaitGroup
-	client           *redisearch.Client
-	insertedDocIds   []string
-	updatedDocIds    []string
-	deletedDocIds    []string
-}
-
-//, client* redisearch.Client,  pipelineSize int, documents []redisearch.Document
-func rowToRSDocument(row string) (document *redisearch.Document) {
-	if debug > 0 {
-		fmt.Fprintln(os.Stderr, "converting row to rediSearch Document "+row)
-	}
-	fieldSizesStr := strings.Split(row, ",")
-	// we need at least the id and score
-	if len(fieldSizesStr) >= 2 {
-		documentId := loader.DatabaseName() + "-" + fieldSizesStr[0]
-		documentScore, _ := strconv.ParseFloat(fieldSizesStr[1], 64)
-		doc := redisearch.NewDocument(documentId, float32(documentScore))
-
-		for _, keyValuePair := range fieldSizesStr[2:] {
-			pair := strings.Split(keyValuePair, "=")
-			if len(pair) == 2 {
-				if debug > 0 {
-					fmt.Fprintln(os.Stderr, "On doc "+documentId+" adding field with NAME "+pair[0]+" and VALUE "+pair[1])
-				}
-				doc.Set(pair[0], pair[1])
-			} else {
-				if debug > 0 {
-					fmt.Fprintf(os.Stderr, "On doc "+documentId+" len(pair)=%d", len(pair))
-				}
-				log.Fatalf("keyValuePair pair size != 2 . Got " + keyValuePair)
-			}
-		}
-		if debug > 0 {
-			fmt.Fprintln(os.Stderr, "Doc "+documentId)
-		}
-		return &doc
-	}
-	return document
-}
-
-func connectionProcessor(p *processor, pipeline uint64, updateRate float64, deleteRate float64, noSaveOption bool, updatePartial bool, updateCondition string) {
-	var documents = make([]redisearch.Document, 0)
-
-	pipelinePos := uint64(0)
-	insertCount := uint64(0)
-	totalBytes := uint64(0)
-	// using random between [0,1) to determine whether it is an delete,update, or insert
-	// DELETE IF BETWEEN [0,deleteLimit)
-	// UPDATE IF BETWEEN [deleteLimit,updateLimit)
-	// INSERT IF BETWEEN [updateLimit,1)
-
-	deleteUpperLimit := 0.0
-	updateUpperLimit := deleteUpperLimit + updateRate
-
-	updateOpts := redisearch.IndexingOptions{
-		Language:         "",
-		NoSave:           noSaveOption,
-		Replace:          true,
-		Partial:          updatePartial,
-		ReplaceCondition: updateCondition,
-	}
-
-	indexingOpts := redisearch.DefaultIndexingOptions
-	indexingOpts.NoSave = noSaveOption
-
-	for row := range p.rows {
-		doc := rowToRSDocument(row)
-		if doc != nil {
-			documentPayload := uint64((*doc).EstimateSize())
-			totalBytes += documentPayload
-			//fmt.Println(totalBytes)
-			(*doc).EstimateSize()
-			val := rand.Float64()
-			// DELETE
-			// TODO:
-			// UPDATE
-			// only possible if we already have something to update
-			if val >= deleteUpperLimit && val < updateUpperLimit && (len(p.insertedDocIds) > 0) {
-				p.insertedDocIds = append(p.insertedDocIds, doc.Id)
-				idToUdpdate := p.insertedDocIds[rand.Intn(len(p.insertedDocIds))]
-				doc.Id = idToUdpdate
-				// make sure we flush the pipeline prior than updating
-				if pipelinePos > 0 {
-					// Index the document. The API accepts multiple documents at a time
-					CommonIndexInsertDocuments(p, indexingOpts, documents, totalBytes, insertCount)
-					documents, insertCount, pipelinePos, totalBytes = LocalCountersReset()
-				}
-				CommonIndexUpdateDocument(p, updateOpts, doc, totalBytes)
-				documents, insertCount, pipelinePos, totalBytes = LocalCountersReset()
-				// INSERT
-			} else {
-				documents = append(documents, *doc)
-				p.insertedDocIds = append(p.insertedDocIds, doc.Id)
-				insertCount++
-				pipelinePos++
-			}
-			if pipelinePos%pipeline == 0 && len(documents) > 0 {
-				// Index the document. The API accepts multiple documents at a time
-				CommonIndexInsertDocuments(p, indexingOpts, documents, totalBytes, insertCount)
-				documents, insertCount, pipelinePos, totalBytes = LocalCountersReset()
-			}
-		}
-
-	}
-	// In the there are still documents to be processed
-	if pipelinePos != 0 && len(documents) > 0 {
-		// Index the document. The API accepts multiple documents at a time
-		CommonIndexInsertDocuments(p, indexingOpts, documents, totalBytes, insertCount)
-		documents, insertCount, pipelinePos, totalBytes = LocalCountersReset()
-	}
-	p.wg.Done()
-}
-
-func CommonIndexUpdateDocument(p *processor, updateOpts redisearch.IndexingOptions, doc *redisearch.Document, totalBytes uint64) {
-	start := time.Now()
-	if err := p.client.IndexOptions(updateOpts, *doc); err != nil {
-		log.Fatalf("failed: %s\n", err)
-	}
-	took := uint64(time.Since(start).Milliseconds())
-	p.updatesChan <- 1
-	updateCommonChannels(p, took, totalBytes)
-}
-
-func CommonIndexInsertDocuments(p *processor, opts redisearch.IndexingOptions, documents []redisearch.Document, bytesCount uint64, insertCount uint64) () {
-	start := time.Now()
-	if err := p.client.IndexOptions(opts, documents...); err != nil {
-		log.Fatalf("failed: %s\n", err)
-	}
-	took := uint64(time.Since(start).Milliseconds())
-	p.insertsChan <- insertCount
-	updateCommonChannels(p, took, bytesCount)
-}
-
-func updateCommonChannels(p *processor, took uint64, bytesCount uint64) {
-	p.totalLatencyChan <- took
-	p.totalBytesChan <- bytesCount
 }
 
 func LocalCountersReset() (documents []redisearch.Document, pipelinePos uint64, insertCount uint64, totalBytes uint64) {
@@ -275,68 +127,6 @@ func LocalCountersReset() (documents []redisearch.Document, pipelinePos uint64, 
 	insertCount = 0
 	totalBytes = 0
 	return documents, insertCount, pipelinePos, totalBytes
-}
-
-func (p *processor) Init(_ int, _ bool) {
-	p.client = redisearch.NewClient(host, loader.DatabaseName())
-}
-
-// ProcessBatch reads eventsBatches which contain rows of data for FT.ADD redis command string
-func (p *processor) ProcessBatch(b load.Batch, doLoad bool, updateRate, deleteRate float64) (uint64, uint64, uint64, uint64, uint64, uint64) {
-	events := b.(*eventsBatch)
-	rowCnt := uint64(len(events.rows))
-	metricCnt := uint64(0)
-	updateCount := uint64(0)
-	deleteCount := uint64(0)
-	totalLatency := uint64(0)
-	totalBytes := uint64(0)
-	if doLoad {
-		buflen := rowCnt + 1
-
-		p.insertsChan = make(chan uint64, buflen)
-		p.updatesChan = make(chan uint64, buflen)
-		p.deletesChan = make(chan uint64, buflen)
-		p.totalLatencyChan = make(chan uint64, buflen)
-		p.totalBytesChan = make(chan uint64, buflen)
-
-		p.wg = &sync.WaitGroup{}
-		p.rows = make(chan string, buflen)
-		p.wg.Add(1)
-		go connectionProcessor(p, pipeline, updateRate, deleteRate, noSave, replacePartial, replacePartialCondition)
-		for _, row := range events.rows {
-			p.rows <- row
-		}
-		close(p.rows)
-		p.wg.Wait()
-		close(p.insertsChan)
-		close(p.updatesChan)
-		close(p.deletesChan)
-		close(p.totalLatencyChan)
-		close(p.totalBytesChan)
-
-		for val := range p.insertsChan {
-			metricCnt += val
-		}
-		for val := range p.updatesChan {
-			updateCount += val
-		}
-		for val := range p.deletesChan {
-			deleteCount += val
-		}
-		for val := range p.totalLatencyChan {
-			totalLatency += val
-		}
-		for val := range p.totalBytesChan {
-			totalBytes += val
-		}
-
-	}
-	events.rows = events.rows[:0]
-	ePool.Put(events)
-	return metricCnt, rowCnt, updateCount, deleteCount, totalLatency, totalBytes
-}
-
-func (p *processor) Close(_ bool) {
 }
 
 func main() {
