+++ load/indexing_result.go
@@ -41,13 +41,13 @@ type TestResult struct {
 	MeasuredDeleteRatio float64 `json:"MeasuredDeleteRatio"`
 
 	// Overall Rates
-	OverallAvgOpsRate    float64 `json:"OverallAvgOpsRate"`
-	OverallAvgInsertRate float64 `json:"OverallAvgInsertRate"`
-	OverallAvgUpdateRate float64 `json:"OverallAvgUpdateRate"`
-	OverallAvgDeleteRate float64 `json:"OverallAvgDeleteRate"`
-	OverallAvgLatency    float64 `json:"OverallAvgLatency"`
-	OverallAvgByteRate   float64 `json:"OverallAvgByteRate"`
-	OverallAvgByteRateHumanReadable   string `json:"OverallAvgByteRateHumanReadable"`
+	OverallAvgOpsRate               float64 `json:"OverallAvgOpsRate"`
+	OverallAvgInsertRate            float64 `json:"OverallAvgInsertRate"`
+	OverallAvgUpdateRate            float64 `json:"OverallAvgUpdateRate"`
+	OverallAvgDeleteRate            float64 `json:"OverallAvgDeleteRate"`
+	OverallAvgLatency               float64 `json:"OverallAvgLatency"`
+	OverallAvgByteRate              float64 `json:"OverallAvgByteRate"`
+	OverallAvgByteRateHumanReadable string  `json:"OverallAvgByteRateHumanReadable"`
 
 	// Time-Series
 	OverallIngestionRateTs  []DataPoint `json:"OverallIngestionRateTs"`
